Извлёк в отдельный модуль сетевое взаимодействие
