--- app/src/main/java/ru/iteco/fmhandroid/api/AuthApi.kt
@@ -1,14 +1,0 @@
-package ru.iteco.fmhandroid.api
-
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
-import ru.iteco.fmh.model.AuthState
-import ru.iteco.fmhandroid.dto.LoginData
-
-interface AuthApi {
-    @POST("authentication/login")
-    suspend fun getTokens(
-        @Body loginData: LoginData
-    ): Response<AuthState>
-}
