--- app/src/main/java/ru/iteco/fmhandroid/api/NewsApi.kt
@@ -1,19 +1,0 @@
-package ru.iteco.fmhandroid.api
-
-import retrofit2.Response
-import retrofit2.http.*
-import ru.iteco.fmhandroid.dto.News
-
-interface NewsApi {
-    @GET("news")
-    suspend fun getAllNews(): Response<List<News>>
-
-    @PUT("news")
-    suspend fun editNewsItem(@Body newsItem: News): Response<News>
-
-    @POST("news")
-    suspend fun saveNewsItem(@Body newsItem: News): Response<News>
-
-    @DELETE("news/{id}")
-    suspend fun removeNewsItemById(@Path("id") id: Int): Response<Unit>
-}
