--- app/src/main/java/ru/iteco/fmhandroid/api/RefreshTokensApi.kt
@@ -1,16 +1,0 @@
-package ru.iteco.fmhandroid.api
-
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.Header
-import retrofit2.http.POST
-import ru.iteco.fmh.model.AuthState
-import ru.iteco.fmhandroid.dto.RefreshRequest
-
-interface RefreshTokensApi {
-    @POST("authentication/refresh")
-    suspend fun refreshTokens(
-        @Header("Authorization") refreshToken: String,
-        @Body refreshRequest: RefreshRequest
-    ): Response<AuthState>
-}
