--- app/src/main/java/ru/iteco/fmhandroid/dto/JwtResponse.kt
@@ -1,6 +1,0 @@
-package ru.iteco.fmhandroid.dto
-
-data class JwtResponse(
-    val code: String,
-    val message: String
-)
