--- app/src/main/java/ru/iteco/fmhandroid/dto/LoginData.kt
@@ -1,6 +1,0 @@
-package ru.iteco.fmhandroid.dto
-
-data class LoginData(
-    val login: String,
-    val password: String
-)
