--- app/src/main/java/ru/iteco/fmhandroid/dto/RefreshRequest.kt
@@ -1,5 +1,0 @@
-package ru.iteco.fmhandroid.dto
-
-data class RefreshRequest(
-    val refreshToken: String
-)
