+++ data/api/src/main/kotlin/ru/iteco/fmh/data/api/AuthApi.kt
@@ -1,0 +1,14 @@
+package ru.iteco.fmh.data.api
+
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.POST
+import ru.iteco.fmh.data.api.dto.LoginData
+import ru.iteco.fmh.model.AuthState
+
+interface AuthApi {
+    @POST("authentication/login")
+    suspend fun getTokens(
+        @Body loginData: LoginData
+    ): Response<AuthState>
+}
