+++ data/api/src/main/kotlin/ru/iteco/fmh/data/api/RefreshTokensApi.kt
@@ -1,0 +1,16 @@
+package ru.iteco.fmh.data.api
+
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.Header
+import retrofit2.http.POST
+import ru.iteco.fmh.data.api.dto.RefreshRequest
+import ru.iteco.fmh.model.AuthState
+
+interface RefreshTokensApi {
+    @POST("authentication/refresh")
+    suspend fun refreshTokens(
+        @Header("Authorization") refreshToken: String,
+        @Body refreshRequest: RefreshRequest
+    ): Response<AuthState>
+}
