+++ data/api/src/main/kotlin/ru/iteco/fmh/data/api/dto/JwtResponse.kt
@@ -1,0 +1,6 @@
+package ru.iteco.fmh.data.api.dto
+
+data class JwtResponse(
+    val code: String,
+    val message: String
+)
