+++ data/api/src/main/kotlin/ru/iteco/fmh/data/api/dto/LoginData.kt
@@ -1,0 +1,6 @@
+package ru.iteco.fmh.data.api.dto
+
+data class LoginData(
+    val login: String,
+    val password: String
+)
