+++ data/api/src/main/kotlin/ru/iteco/fmh/data/api/dto/RefreshRequest.kt
@@ -1,0 +1,5 @@
+package ru.iteco.fmh.data.api.dto
+
+data class RefreshRequest(
+    val refreshToken: String
+)
